Preserve BC for users who use pull requests + includes
This commit further revises the design in the previous two commits.

Broadly, the issue with the previous design is that it still breaks
compatibility. In the example of a pipeline given:

  ---
  kind: "pipeline"
  type: "kubernetes"
  name: "exclude-open"
  steps:
    - name: "container"
      image: "busybox"
      commands:
        - echo "Pushed"
  trigger:
    event:
      - pull_request
    action:
      exclude:
        - opened

The code is configured to opt-in to the "close" event. However, that
breaks existing users who might be using the pipeline currently just
desiring the synchronize event.

The desire is not to break these users. Given this, the logic has been
made safer still -- it requires users to deliberately opt in to the close
events with "include". As a nice side effect, the logic of this is now
simpler.

== Design notes
=== Non-consistent behaviour

Naively the author considered this would be "non-consistent" or
"non-reasonable" behavior. However, preserving BC is a far more
important property than the inherent reasonability of this behaviour.

=== Tests

Tests turned out to be tremendously useful here.